project/requirements.txt
adjacent==1.1.0
backports.shutil-get-terminal-size==1.0.0
boto3==1.9.23
botocore==1.12.23
cent==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.5.1
decorator==4.2.1
Django==1.11.11
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-json-rpc==0.7.2
django-storages==1.7.1
docutils==0.14
enum34==1.1.6
factory-boy==2.11.1
Faker==0.9.1
funcsigs==1.0.2
futures==3.2.0
idna==2.7
ipaddress==1.0.22
ipython==5.5.0
ipython-genutils==0.2.0
jmespath==0.9.3
mock==2.0.0
mysqlclient==1.3.13
oauthlib==2.1.0
pathlib2==2.3.0
pbr==4.3.0
pexpect==4.4.0
pickleshare==0.7.4
pip-chill==0.1.7
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.4
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-openid==2.2.5
pytz==2018.3
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
